pyproject.toml
[project]
name = "entropick"
version = "0.1.0"
description = "Fair and random reviewer assignment for pull requests via Slack app."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
]

# Ruff configuration - minimal setup
[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
# Basic linting rules
select = ["E", "F", "W", "I"]  # pycodestyle errors/warnings, Pyflakes, isort

[tool.ruff.format]
quote-style = "double"

# Pyright type checking configuration - basic setup
[tool.pyright]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pyright>=1.1.401",
    "ruff>=0.11.11",
]
